video_player

+++ src/rtsp/ffmpeg.py
@@ -48,14 +48,16 @@ async def FFmpegRead(url: str, height: int, width: int, gpu: int, websocket: Web
                     if not result:
                         logger.error("OpenCV encoding error")
                         break
-
+                    
                     await websocket.send_bytes(encimg.tobytes())
                     logger.debug(f"FFmpegRead | Sent {len(encimg.tobytes())} bytes of data.")
         except Exception as e:
             logger.error(f"Error in FFmpegRead: {e}")
         finally:
-            process.stdout.close()
-            await process.terminate()
-            logger.info("FFmpegRead | FFmpeg read process has completed.")
+            if process and process.returncode is None:
+                process.stdout.close()
+                await process.terminate()
+                await process.wait()
+                logger.info("FFmpegRead | FFmpeg read process has completed.")
         await asyncio.sleep(5)
         logger.info("FFmpegRead | Restarting FFmpeg read process")
